+++ services/test/mock/mock_accessibility_element_operator_proxy.cpp
@@ -33,7 +33,7 @@ bool MockAccessibilityElementOperatorProxy::WriteInterfaceToken(MessageParcel& d
     return true;
 }
 
-void MockAccessibilityElementOperatorProxy::SearchElementInfoByAccessibilityId(const int64_t elementId,
+void MockAccessibilityElementOperatorProxy::SearchElementInfoByAccessibilityId(const int32_t elementId,
     const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback>& callback, const int32_t mode)
 {
     GTEST_LOG_(INFO) << "MockAccessibilityElementOperatorProxy SearchElementInfoByAccessibilityId";
@@ -44,7 +44,7 @@ void MockAccessibilityElementOperatorProxy::SearchElementInfoByAccessibilityId(c
     return;
 }
 
-void MockAccessibilityElementOperatorProxy::SearchElementInfosByText(const int64_t elementId, const std::string& text,
+void MockAccessibilityElementOperatorProxy::SearchElementInfosByText(const int32_t elementId, const std::string& text,
     const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback>& callback)
 {
     GTEST_LOG_(INFO) << "MockAccessibilityElementOperatorProxy SearchElementInfosByText";
@@ -55,7 +55,7 @@ void MockAccessibilityElementOperatorProxy::SearchElementInfosByText(const int64
     return;
 }
 
-void MockAccessibilityElementOperatorProxy::FindFocusedElementInfo(const int64_t elementId, const int32_t focusType,
+void MockAccessibilityElementOperatorProxy::FindFocusedElementInfo(const int32_t elementId, const int32_t focusType,
     const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback>& callback)
 {
     GTEST_LOG_(INFO) << "MockAccessibilityElementOperatorProxy FindFocusedElementInfo";
@@ -66,7 +66,7 @@ void MockAccessibilityElementOperatorProxy::FindFocusedElementInfo(const int64_t
     return;
 }
 
-void MockAccessibilityElementOperatorProxy::FocusMoveSearch(const int64_t elementId, const int32_t direction,
+void MockAccessibilityElementOperatorProxy::FocusMoveSearch(const int32_t elementId, const int32_t direction,
     const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback>& callback)
 {
     GTEST_LOG_(INFO) << "MockAccessibilityElementOperatorProxy FocusMoveSearch";
@@ -77,7 +77,7 @@ void MockAccessibilityElementOperatorProxy::FocusMoveSearch(const int64_t elemen
     return;
 }
 
-void MockAccessibilityElementOperatorProxy::ExecuteAction(const int64_t elementId, const int32_t action,
+void MockAccessibilityElementOperatorProxy::ExecuteAction(const int32_t elementId, const int32_t action,
     const std::map<std::string, std::string> &actionArguments, const int32_t requestId,
     const sptr<IAccessibilityElementOperatorCallback>& callback)
 {
